Move editing to SavedPaymentMethodMutator.
